Moved the shared logic of the AzureKeyVaultCommand and CertificateStoreCommand to a new Handle method inside the CodeCommand
