+++ src/Sign.Cli/CertificateStoreCommand.cs
@@ -6,11 +6,6 @@ using System.CommandLine;
 using System.CommandLine.Invocation;
 using System.CommandLine.IO;
 using System.Security.Cryptography;
-using System.Text;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.FileSystemGlobbing;
-using Microsoft.Extensions.FileSystemGlobbing.Abstractions;
-using Microsoft.Extensions.Logging;
 using Sign.Core;
 using Sign.SignatureProviders.CertificateStore;
 
@@ -53,33 +48,6 @@ namespace Sign.Cli
 
             this.SetHandler(async (InvocationContext context) =>
             {
-                DirectoryInfo baseDirectory = context.ParseResult.GetValueForOption(_codeCommand.BaseDirectoryOption)!;
-                string? applicationName = context.ParseResult.GetValueForOption(_codeCommand.ApplicationNameOption);
-                string? publisherName = context.ParseResult.GetValueForOption(_codeCommand.PublisherNameOption);
-                string? description = context.ParseResult.GetValueForOption(_codeCommand.DescriptionOption);
-                // This option is required.  If its value fails to parse we won't have reached here,
-                // and after successful parsing its value will never be null.
-                // Non-null is already guaranteed; the null-forgiving operator (!) just simplifies code.
-                Uri descriptionUrl = context.ParseResult.GetValueForOption(_codeCommand.DescriptionUrlOption)!;
-                string? fileListFilePath = context.ParseResult.GetValueForOption(_codeCommand.FileListOption);
-                HashAlgorithmName fileHashAlgorithmName = context.ParseResult.GetValueForOption(_codeCommand.FileDigestOption);
-                HashAlgorithmName timestampHashAlgorithmName = context.ParseResult.GetValueForOption(_codeCommand.TimestampDigestOption);
-                // This option is optional but has a default value.  If its value fails to parse we won't have
-                // reached here, and after successful parsing its value will never be null.
-                // Non-null is already guaranteed; the null-forgiving operator (!) just simplifies code.
-                Uri timestampUrl = context.ParseResult.GetValueForOption(_codeCommand.TimestampUrlOption)!;
-                LogLevel verbosity = context.ParseResult.GetValueForOption(_codeCommand.VerbosityOption);
-                string? output = context.ParseResult.GetValueForOption(_codeCommand.OutputOption);
-                int maxConcurrency = context.ParseResult.GetValueForOption(_codeCommand.MaxConcurrencyOption);
-
-                string? certificateFingerprint = context.ParseResult.GetValueForOption(CertificateFingerprintOption);
-                HashAlgorithmName certificateFingerprintAlgorithm = context.ParseResult.GetValueForOption(CertificateFingerprintAlgorithmOption);
-                string? certificatePath = context.ParseResult.GetValueForOption(CertificateFileOption);
-                string? certificatePassword = context.ParseResult.GetValueForOption(CertificatePasswordOption);
-                string? cryptoServiceProvider = context.ParseResult.GetValueForOption(CryptoServiceProviderOption);
-                string? privateKeyContainer = context.ParseResult.GetValueForOption(PrivateKeyContainerOption);
-                bool useMachineKeyContainer = context.ParseResult.GetValueForOption(UseMachineKeyContainerOption);
-
                 string? fileArgument = context.ParseResult.GetValueForArgument(FileArgument);
 
                 if (string.IsNullOrEmpty(fileArgument))
@@ -88,6 +56,16 @@ namespace Sign.Cli
                     context.ExitCode = ExitCode.InvalidOptions;
                     return;
                 }
+
+                // Some of the options are required and that is why we can safely use
+                // the null-forgiving operator (!) to simplify the code.
+                string certificateFingerprint = context.ParseResult.GetValueForOption(CertificateFingerprintOption)!;
+                HashAlgorithmName certificateFingerprintAlgorithm = context.ParseResult.GetValueForOption(CertificateFingerprintAlgorithmOption);
+                string? certificatePath = context.ParseResult.GetValueForOption(CertificateFileOption);
+                string? certificatePassword = context.ParseResult.GetValueForOption(CertificatePasswordOption);
+                string? cryptoServiceProvider = context.ParseResult.GetValueForOption(CryptoServiceProviderOption);
+                string? privateKeyContainer = context.ParseResult.GetValueForOption(PrivateKeyContainerOption);
+                bool useMachineKeyContainer = context.ParseResult.GetValueForOption(UseMachineKeyContainerOption);
 
                 // Certificate Fingerprint is required in case the provided certificate container contains multiple certificates.
                 if (string.IsNullOrEmpty(certificateFingerprint))
@@ -117,140 +95,17 @@ namespace Sign.Cli
                     }
                 }
 
-                // Make sure this is rooted
-                if (!Path.IsPathRooted(baseDirectory.FullName))
-                {
-                    context.Console.Error.WriteFormattedLine(
-                        Resources.InvalidBaseDirectoryValue,
-                        _codeCommand.BaseDirectoryOption);
-                    context.ExitCode = ExitCode.InvalidOptions;
-                    return;
-                }
+                CertificateStoreServiceProvider certificateStoreServiceProvider = new(
+                    certificateFingerprint,
+                    certificateFingerprintAlgorithm,
+                    cryptoServiceProvider,
+                    privateKeyContainer,
+                    certificatePath,
+                    certificatePassword,
+                    useMachineKeyContainer);
 
-                IServiceProvider serviceProvider = serviceProviderFactory.Create(
-                    verbosity,
-                    addServices: (IServiceCollection services) =>
-                    {
-                        CertificateStoreServiceProvider certificateStoreServiceProvider = new(
-                            certificateFingerprint,
-                            certificateFingerprintAlgorithm,
-                            cryptoServiceProvider,
-                            privateKeyContainer,
-                            certificatePath,
-                            certificatePassword,
-                            useMachineKeyContainer);
-
-                        services.AddSingleton<ISignatureAlgorithmProvider>(
-                            (IServiceProvider serviceProvider) => certificateStoreServiceProvider.GetSignatureAlgorithmProvider(serviceProvider));
-                        services.AddSingleton<ICertificateProvider>(
-                            (IServiceProvider serviceProvider) => certificateStoreServiceProvider.GetCertificateProvider(serviceProvider));
-                    });
-
-                List<FileInfo> inputFiles;
-
-                // If we're going to glob, we can't be fully rooted currently (fix me later)
-
-                bool isGlob = fileArgument.Contains('*');
-
-                if (isGlob)
-                {
-                    if (Path.IsPathRooted(fileArgument))
-                    {
-                        context.Console.Error.WriteLine(Resources.InvalidFileValue);
-                        context.ExitCode = ExitCode.InvalidOptions;
-                        return;
-                    }
-
-                    IFileListReader fileListReader = serviceProvider.GetRequiredService<IFileListReader>();
-                    IFileMatcher fileMatcher = serviceProvider.GetRequiredService<IFileMatcher>();
-
-                    using (MemoryStream stream = new(Encoding.UTF8.GetBytes(fileArgument)))
-                    using (StreamReader reader = new(stream))
-                    {
-                        fileListReader.Read(reader, out Matcher? matcher, out Matcher? antiMatcher);
-
-                        DirectoryInfoBase directory = new DirectoryInfoWrapper(baseDirectory);
-
-                        IEnumerable<FileInfo> matches = fileMatcher.EnumerateMatches(directory, matcher);
-
-                        if (antiMatcher is not null)
-                        {
-                            IEnumerable<FileInfo> antiMatches = fileMatcher.EnumerateMatches(directory, antiMatcher);
-                            matches = matches.Except(antiMatches, FileInfoComparer.Instance);
-                        }
-
-                        inputFiles = matches.ToList();
-                    }
-                }
-                else
-                {
-                    inputFiles = new List<FileInfo>()
-                    {
-                        new FileInfo(ExpandFilePath(baseDirectory, fileArgument))
-                    };
-                }
-
-                FileInfo? fileList = null;
-                if (!string.IsNullOrEmpty(fileListFilePath))
-                {
-                    if (Path.IsPathRooted(fileListFilePath))
-                    {
-                        fileList = new FileInfo(fileListFilePath);
-                    }
-                    else
-                    {
-                        fileList = new FileInfo(ExpandFilePath(baseDirectory, fileListFilePath));
-                    }
-                }
-
-                if (inputFiles.Count == 0)
-                {
-                    context.Console.Error.WriteLine(Resources.NoFilesToSign);
-                    context.ExitCode = ExitCode.NoInputsFound;
-                    return;
-                }
-
-                if (inputFiles.Any(file => !file.Exists))
-                {
-                    context.Console.Error.WriteFormattedLine(
-                        Resources.SomeFilesDoNotExist,
-                        _codeCommand.BaseDirectoryOption);
-
-                    foreach (FileInfo file in inputFiles.Where(file => !file.Exists))
-                    {
-                        context.Console.Error.WriteLine($"    {file.FullName}");
-                    }
-
-                    context.ExitCode = ExitCode.NoInputsFound;
-                    return;
-                }
-
-                ISigner signer = serviceProvider.GetRequiredService<ISigner>();
-
-                context.ExitCode = await signer.SignAsync(
-                    inputFiles,
-                    output,
-                    fileList,
-                    baseDirectory,
-                    applicationName,
-                    publisherName,
-                    description,
-                    descriptionUrl,
-                    timestampUrl,
-                    maxConcurrency,
-                    fileHashAlgorithmName,
-                    timestampHashAlgorithmName);
+                await _codeCommand.HandleAsync(context, serviceProviderFactory, certificateStoreServiceProvider, fileArgument);
             });
-        }
-
-        private static string ExpandFilePath(DirectoryInfo baseDirectory, string file)
-        {
-            if (Path.IsPathRooted(file))
-            {
-                return file;
-            }
-
-            return Path.Combine(baseDirectory.FullName, file);
         }
     }
 }
