+++ src/Sign.Core/ISignatureProvider.cs
@@ -1,0 +1,12 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE.txt file in the project root for more information.
+
+namespace Sign.Core
+{
+    internal interface ISignatureProvider
+    {
+        ISignatureAlgorithmProvider GetSignatureAlgorithmProvider(IServiceProvider serviceProvider);
+        ICertificateProvider GetCertificateProvider(IServiceProvider serviceProvider);
+    }
+}
